bot/requirements.txt
aiogram>=3.0.0
aiohttp
aiosqlite
python-dotenv
loguru
psutil
qrcode
pillow
